bugfix: types

+++ src/clearcontrol/microscope/lightsheet/imaging/interleaved/AppendConsecutiveInterleavedImagingInstruction.java
@@ -65,7 +65,7 @@ public class AppendConsecutiveInterleavedImagingInstruction extends LightSheetMi
             index++;
             schedule.add(index, new DropOldestStackInterfaceContainerInstruction(FusedImageDataContainer.class, getLightSheetMicroscope().getDataWarehouse()));
             index++;
-            schedule.add(index, new PauseUntilTimeAfterMeasuredTimeInstruction(timeMeasurementKey, (long)(mIntervalInSeconds * 1000)));
+            schedule.add(index, new PauseUntilTimeAfterMeasuredTimeInstruction(timeMeasurementKey, (int)(mIntervalInSeconds * 1000)));
             index++;
         }
         return true;
